# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DzPulse - Algeria Public Polling Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /polls:
    get:
      operationId: listPolls
      tags: [polls]
      summary: List polls with filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category slug
        - name: status
          in: query
          schema:
            type: string
            enum: [open, closed, upcoming, archived]
        - name: sort
          in: query
          schema:
            type: string
            enum: [trending, latest, most_voted, controversial]
          description: Sort order
        - name: wilaya
          in: query
          schema:
            type: string
          description: Filter by wilaya code or "national"
        - name: search
          in: query
          schema:
            type: string
          description: Search query
        - name: featured
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of polls
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollListResponse"
  /polls/{slug}:
    get:
      operationId: getPollBySlug
      tags: [polls]
      summary: Get poll detail by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Poll detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PollDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /polls/{slug}/related:
    get:
      operationId: getRelatedPolls
      tags: [polls]
      summary: Get related polls
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Related polls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Poll"

  # Votes
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        color:
          type: string
        icon:
          type: string
        isActive:
          type: boolean
        orderIndex:
          type: integer
        pollCount:
          type: integer
      required: [id, name, slug, color, icon, isActive, orderIndex, pollCount]
    PollOption:
      type: object
      properties:
        id:
          type: integer
        label:
          type: string
        slug:
          type: string
        displayOrder:
          type: integer
        voteCount:
          type: integer
        percentageCache:
          type: number
      required: [id, label, slug, displayOrder, voteCount, percentageCache]
    Poll:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        subtitle:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        pollType:
          type: string
          enum: [binary, multiple_choice]
        status:
          type: string
          enum: [draft, open, closed, upcoming, archived]
        regionScope:
          type: string
          enum: [national, wilaya]
        wilayaCode:
          type: string
          nullable: true
        totalVotes:
          type: integer
        uniqueVoters:
          type: integer
        isFeatured:
          type: boolean
        isTrending:
          type: boolean
        opensAt:
          type: string
          format: date-time
          nullable: true
        closesAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        category:
          $ref: "#/components/schemas/Category"
        options:
          type: array
          items:
            $ref: "#/components/schemas/PollOption"
        tags:
          type: array
          items:
            type: string
      required: [id, title, slug, pollType, status, regionScope, totalVotes, uniqueVoters, isFeatured, isTrending, createdAt, updatedAt, category, options, tags]
    PollDetail:
      allOf:
        - $ref: "#/components/schemas/Poll"
        - type: object
          properties:
            editorialNote:
              type: string
              nullable: true
            methodologyNote:
              type: string
              nullable: true
            sourceLinks:
              type: array
              items:
                type: string
            commentCount:
              type: integer
          required: [commentCount]
    PollListResponse:
      type: object
      properties:
        polls:
          type: array
          items:
            $ref: "#/components/schemas/Poll"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [polls, total, page, limit, totalPages]
...